add util.unzip_granule, unzip if needed in rtc_sentinel cli

+++ hyp3_gamma/util.py
@@ -19,11 +19,17 @@ def find_and_remove(directory, file_pattern):
 def get_granule(granule):
     download_url = get_download_url(granule)
     zip_file = download_file(download_url, chunk_size=10485760)
-    log.info(f'Unzipping {zip_file}')
+    safe_dir = unzip_granule(zip_file, remove=True)
+    return safe_dir
+
+
+def unzip_granule(zip_file, remove=False):
     with ZipFile(zip_file) as z:
         z.extractall()
-    os.remove(zip_file)
-    return f'{granule}.SAFE'
+        safe_dir = next(item.filename for item in z.infolist() if item.is_dir() and item.filename.endswith('.SAFE/'))
+    if remove:
+        os.remove(zip_file)
+    return safe_dir
 
 
 def earlier_granule_first(g1, g2):
